third_party/blink/web_tests/fast/css/webkit-mask-crash-implicit-expected.txt

PASS document.styleSheets[1].rules[0].style.cssText is "mask-repeat: repeat-x;"
PASS document.styleSheets[1].rules[0].style.getPropertyValue('-webkit-mask') is ""
PASS document.styleSheets[1].rules[0].style.getPropertyValue('mask-repeat') is "repeat-x"
PASS successfullyParsed is true

TEST COMPLETE
